“fit_fixed_bezier” fixed

+++ make_error.py
@@ -17,7 +17,7 @@ def main():
     parser.add_argument('--num_steps', type=int, default=100, help='Number of steps for the fitting process.')
     parser.add_argument('--input_file', type=str, default='examples/g-150.txt', help='Path to the input file containing points.')
     parser.add_argument('--output_error', type=str, default='errors/error_g.npy', help='Path to save the output error.')
-    parser.add_argument('--epsilon', type=int, default=2, help='Epsilon value for the RDP algorithm.')
+    parser.add_argument('--epsilon', type=float, default=2, help='Epsilon value for the RDP algorithm.')
 
     args = parser.parse_args()
